# Reserve capacity before extending Punctuated