--- a/tests/test_iterators.rs
+++ b/tests/test_iterators.rs
@@ -68,3 +68,31 @@ fn may_dangle() {
         }
     }
 }
+
+#[test]
+fn extend() {
+    let mut p: Punctuated<i32, Token![,]> = punctuated!(1, 2, 3);
+
+    // No punctuation at the end after calling extend on an empty iterator of items T.
+    p.extend(None::<i32>);
+    assert_eq!(p, punctuated!(1, 2, 3));
+
+    // But punctuation at the end after calling extend on an empty iterator of pairs.
+    // This behavior is not consistent and should be changed in a future breaking release
+    // to behave like above.
+    p.extend(None::<Pair<_, _>>);
+    let mut expected = punctuated!(1, 2, 3);
+    expected.push_punct(Default::default());
+    assert_eq!(p, expected);
+
+    p.extend([4, 5, 6]);
+    assert_eq!(p, punctuated!(1, 2, 3, 4, 5, 6));
+
+    p.extend([Pair::Punctuated(7, Default::default())]);
+    let mut expected = punctuated!(1, 2, 3, 4, 5, 6, 7);
+    expected.push_punct(Default::default());
+    assert_eq!(p, expected);
+
+    p.extend([Pair::Punctuated(8, Default::default()), Pair::End(9)]);
+    assert_eq!(p, punctuated!(1, 2, 3, 4, 5, 6, 7, 8, 9));
+}