While extending `Punctuated` from an iterator, use `size_hint` to reserve capacity before pushing to avoid repeated allocations.

I did also avoid unneeded allocations of a new `Box`.

There are also small optimizations like removing the check of `nomore` in the while loop and only checking if we did receive an `End`.

I did mention in a TODO comment that the behavior of `extend` from an iterator of pairs is not consistent with the behavior of `extend` from an iterator of items. You can also see it in the added test. I would recommend not adding a punctuation if the iterator is empty, but since that could be a braking change, I did not do it.

This is my first contribution and I did not find a CONTRIBUTING file. Therefore, please tell me if something is odd :)

Sorry for the typo in the branch name :/