@mo8it 2023-06-24
I have done the following benchmark:

`cargo b -r && hyperfine -N "BINARY"`:

```rust
for _ in 0..10_000 {
    std::hint::black_box(Punctuated::<u64, syn::token::Comma>::from_iter(0..4096));
    std::hint::black_box(Punctuated::<u64, syn::token::Comma>::from_iter(
        (0..4096).map(|i| syn::punctuated::Pair::Punctuated(i, Default::default())),
    ));
}
```

Results:

`master` branch:

```
Time (mean ± σ):     630.4 ms ±   9.4 ms    [User: 628.7 ms, System: 0.7 ms]
Range (min … max):   617.5 ms … 642.1 ms    10 runs
```

This branch:

```
Time (mean ± σ):     257.0 ms ±   4.3 ms    [User: 256.0 ms, System: 0.5 ms]
Range (min … max):   253.3 ms … 265.7 ms    11 runs
```

The speed up for this benchmark is about 2.5x :D