@mo8it 2023-06-24
I think that the assertions in `push_punct` and `push_value` should be `debug_assert!` instead of `assert!`. People using them are opting in for better performance, otherwise they would just use `push`.