@mo8it 2023-06-24
While waiting for a review, I did just see that I could recycle `Box` in `push` too!

Here is a benchmark:

```rust
for _ in 0..1_000_000 {
    let mut p: Punctuated<_, syn::token::Comma> = Punctuated::new();
    for i in 0..16 {
        p.push(i);
    }
}
```

`master` branch:

```
Time (mean ± σ):     240.9 ms ±   1.7 ms    [User: 240.0 ms, System: 0.4 ms]
Range (min … max):   239.2 ms … 245.3 ms    12 runs
```

This branch:

```
Time (mean ± σ):     124.2 ms ±   2.2 ms    [User: 123.5 ms, System: 0.4 ms]
Range (min … max):   121.1 ms … 127.1 ms    23 runs
```

Speedup about 1.9x