@mo8it 2023-06-25
Actually, why is `last` boxed in the first place?