@mo8it 2023-07-24
@dtolnay I did read some of syn's source code while working on [typed-builder](https://github.com/idanarye/rust-typed-builder) and did find out that it can be optimized to avoid allocations.

I did just remove some of `unsafe` to reduce the complexity. I did declare that function `unsafe` because I wanted to use `unwrap_unchecked`, but this is not compatible with the current MSRV.

Currently, `unsafe` is only used once to avoid the comparison of the length and capacity before pushing because we do the allocation. I can remove it if you prefer that and just use `Vec::push`.

More complexity could be avoided if `last` is not boxed. Could you please explain to me why it is boxed? Probably to avoid having huge objects in the stack, but is this really the case here?

I think that 2.5x speedup is worth it, but this is only my opinion.

I wanted to add `reserve` and `with_capacity` later to further optimize pushing. But that would be another PR.