@arendjr 2023-07-24
As a heavy user of `typed-builder` and also the author of [`fp-bindgen`](https://github.com/fiberplane/fp-bindgen), both of which depend heavily on `syn`, I am very much interested in work that speeds up compilation time of `syn`-using macros.

Not sure how much this PR is going to help on our overall build times of course, but I might give it a shot soon to see if this PR improves build times for us.