@arendjr 2023-07-25
I tried recompiling our crates that heavily use both `typed-builder` and `fp-bindgen`, both with `syn 2.0.20` and with this branch, but I have to admit the results appear to be a wash. Sometimes one was faster, sometimes the other. In any case they were within margin of error from each other.