@mo8it 2023-07-27
@dtolnay Alright, thank you a lot for this explanation.

I am still interested in knowing why `last` is boxed, just out of curiosity.

Would it be fine for you if I would at least add `with_capacity` and `reserve`? They don't make the code more complex, just 2 simple additional methods. I am fine with it if you say that you don't want it.